Chore(.auth): separate mock login api
